Base crank-gauge show/hide on crank and time
Previously the show/hide of the gauge was based purely on crank
position, and fast cranking would cause the animation be be
unregisterable.

Instead, initiate the gauge hide/show animation based on crank position,
but then progress the animation based on time, to ensure that the
change is displayed on a human visible scale.