Improve sprite draw performance.
Adding tile backgrounds causes on-device performance to plummet to ~10
fps when moving the viewport around screen.

Disables dirty rect tracking and draws all sprites on every frame. The
number of sprites onscreen is large, and the performance of dirty rect
tracking deteriorates quickly while drawing on every frame remains
performant.

Also adds a helper for creating an update-less and collision-less
sprite, which is expected to be the overwhelming majority of sprites.