Split crank-time into crank and game-clock
Crank drives game clock tick rate, but game clock is no longer directly
driven by crank, so it makes sense to split the two.

This will allow other modules to listen to the crank directly, rather
than to game time, which will be useful for displaying time flow gauge
ui and additional menus that may appear when the game is paused.