Add helper script to sanitize game assets.
Piskel is nice to use, but it adds small deviations to pixel RGBA values
for certain operations (fill seems to be problematic). If a pixel has a
tiny fractioin of alpha, it will be rendered by the Playdate as opaque.

Ensure that all partial alpha values are bucketed to either 0 or 1.