Convert game over to Drifter
The resources required for an ant colony simulation were looking to push
the boundaries of the Playdate more and more with every change.

The goal was to of have at least 100 ants,  a reasonable map world size of 200
tiles (~ 1 screen wide, 3 screens high), a varied ecosystem of
cohabitants, food/resource management, etc. A rough estimate of
~500-1000 total entities.

Despite continual efforts to boost performance, the random simulation of
~800 very basic ant behaviors per frame are about the limit before the
game drops below 30 fps. In the future there will be fewer ants,
but on average the logic behind each entity is expected to be much
higher anyway (the ant "ai" is a very short function).

Rather than continuing to fight what seems to be a losing battle, pivot
to a different concept that requires far fewer entities. Currently the
idea is to reuse almost all of the game code to tell a story about an
old frontier town.