Add camera object
Allows the viewport to be either fixed in place, controllable with the
d-pad, or tracking on an entity within the world.

Most gameplay is expected to run in the tracking mode, where the camera
follows the main character. The dpad mode is expected to be useful while
creating the game world during development.