Update world to be based on 32x32 tile size
Previously was 40x40, because that fit the ant sprite at the origin of
this project better. 32x32 will hopefully make texturing easier.

Also moves utils/random to utils/math, because it seemed relevantly
useful at the time, and is a nice place to put general math helpers in
the future.